Fix: #10 Added truncating addition Date and Month
An UnsupportedTemporalType exception would appear
when adding to a unit of time not supported
by periodtype, occurring in time:plus. Support was
included to allow all additions. The additions
must update the relevant units or face truncation.
For example, 24 hours added to 01-01 changes it
to 01-02, where as 23 hours does not.